[#1]refactor: DataSource Refactor 초기 작업
DataModel 이 자기 자신을 관리하는 구조로 만들기 위해 아래와 같은 작업을 실시

- JSONDataManger 를 TaskDataSource 로 대체
- JSONDataMangeable 를 TaskTableViewDataSource 로 대체
